Introduce cron to schedule the metric collection.
Scheduling the job to collect the metrics in the specified intervals and stores the latest available value in memory.
This is to reduce the load on the gocd server when api requests are made to go
more importantly to reduce resource consumed while computing the pipeline disk size.

All expressions supported by https://github.com/robfig/cron will be supported